training/requirements.txt
evalutils==0.2.4
scikit-learn==0.24.2
scipy ==1.4.1
SimpleITK==2.0.2
monai==0.5.0
numpy==1.18.3
fill-voids==2.0.2
pydicom==2.1.2
tqdm==4.47.0
scikit-image==0.17.1
nibabel==3.2.0
pyyaml==6.0